Implemented viewing of launchers and uploading of launchers
Also generated db.xml.gz with the launchers info included. I also
fixed bugs which had the program calculating the wrong filesize
(And thus uploading the wrong filesize too).

+++ src-plugins/Fiji_Updater/src/fiji/pluginManager/logic/PluginDownload.java
@@ -22,7 +22,7 @@ public class PluginDownload implements FileDownload {
 		return plugin;
 	}
 
-	public int getRecordedFileSize() {
+	public long getRecordedFileSize() {
 		return plugin.getFilesize(); //always return the latest version's size
 	}
